% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories.R
\name{AddSuffix}
\alias{AddSuffix}
\alias{AddPreffix}
\title{Function factories for adding preffixes or Suffixes}
\usage{
AddSuffix(suffix = "")

AddPreffix(preffix = "")
}
\arguments{
\item{suffix, preffix}{character that will be pasted pefore or after the string}
}
\description{
Function factories for adding preffixes or Suffixes
}
