% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kableExtra.R
\name{columns_spec}
\alias{columns_spec}
\alias{spec_color}
\alias{kbl}
\title{Helpers for kableExtra}
\usage{
columns_spec(kable_input, columns, ...)

spec_color(
  x,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  na_color = "#BBBBBB",
  scale_from = NULL
)

kbl(data, ...)
}
\arguments{
\item{kable_input}{a kable object created by [shceof::kbl()] (not [kableExtra::kbl()]).}

\item{columns}{a vector indicating which column(s) to be selected.}

\item{...}{other arguments to [kableExtra::column_spec] with support for
special symbols. See Details.


`spec_color` works the same as [kableExtra::spec_color()] with the difference
that the `option` argument can be a palette function returned by
`grDevices::colorRampPalette()`.

For `columns_spec()` (notice the s), the arguments in `...` can use two
especial symbols. `.table` will refer to the original data.frame object,
and `.col` will refer to each column. The arguments will be evaluated once per
column. This all means that if you have a function `colour_function` that takes
numeric argument and returns assigns a colour to each element, then
`... %>% columns_spec(1:2, colour = colour_function(.table[[.col]])` will
colourise the first and second column.

For `columns_spec()` to work, the table must be generated with `shceof::kbl()`, not
`kableExtra::kbl()`. Also bear in mind that some kableExtra operations will remove
the formating necesary for it to work, so just to be safe, use `columns_spec()` right after `kbl()`.}

\item{x}{continuous vectors of values}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the viridis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the viridis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{option}{A character string indicating the colormap option to use. Four
options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"),
"viridis" (or "D", the default option) and "cividis" (or "E").}

\item{na_color}{color code for NA values}

\item{scale_from}{input range (vector of length two). If not given,
is calculated from the range of x}

\item{data}{a data.frame-like object.}
}
\description{
Replacements for kableExtra functions that make it easier to directly pipe
resutls
}
