% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-range.R
\name{common_range}
\alias{common_range}
\title{Get the overlapping range of grouped observations.}
\usage{
common_range(x, groups)
}
\arguments{
\item{x}{vector of observations}

\item{groups}{vector of groups}
}
\description{
Get the overlapping range of grouped observations.
}
