% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Compute correlations}
\usage{
correlate(x, y, signif = 2)
}
\arguments{
\item{x, y}{numeric variables to correlate.}

\item{signif}{number of significant figures to keep in the reporting text.}
}
\description{
Thin wrapper around [cor.test] which returns estimate, p.value, low and high
95% confidence interval and a text ready for reporting.
}
