% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_locations.R
\name{data_path}
\alias{data_path}
\alias{ERA5_TOC}
\alias{ERA5_geopotential}
\alias{ERA5_geopotential_all}
\alias{ERA5_temperature}
\alias{ERA5_vorticity}
\alias{ERA5_ozone}
\alias{ERA5_BE}
\alias{ERA5_T2M}
\alias{NCEP_PSI}
\alias{CMAP}
\alias{ERSST}
\alias{ENSO}
\alias{DMI}
\alias{SAM}
\title{Data locations}
\usage{
data_path(type = c("raw", "derived"), ...)

ERA5_TOC(from_source = FALSE)

ERA5_geopotential(from_source = FALSE)

ERA5_geopotential_all(from_source = FALSE)

ERA5_temperature(from_source = FALSE)

ERA5_vorticity(from_source = FALSE)

ERA5_ozone(from_source = FALSE)

ERA5_BE(from_source = FALSE)

ERA5_T2M(from_source = FALSE)

NCEP_PSI(from_source = FALSE)

CMAP(from_source = FALSE)

ERSST(from_source = FALSE)

ENSO(from_source = FALSE)

DMI(from_source = FALSE)

SAM(from_source = FALSE)
}
\arguments{
\item{type}{character indicating the type of data. The resulting base will be type_data}

\item{...}{characters to be concatenated with [file.path()]}

\item{from_source}{Logical indicating whether to force download of data from the
remote source instead of the Zeonodo repository.}
}
\description{
Functions that return the location of different datasets. They check that
relevant file exist and then return it with a class that has a neat print method.
}
\details{
Each dataset has two potential sources, either the original source or the
Zenodo repository associated with this paper. The default behaviour is to
download the file from Zenodo if the file does not exist. This will
get the exact files that where used for the paper.
If `from_source = TRUE`, then the function will fetch the data from the
primary source (e.g. Climate Data Store) and save the file where it should be.

In either case, the file hash will be compared with a reference hash and warn
if there are differences.
}
