% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{fftspectrum}
\alias{fftspectrum}
\title{FFT spectrum}
\usage{
fftspectrum(x, spans = NULL, B = 10000, ..., probs = 0.95)
}
\arguments{
\item{x}{numeric vector}

\item{spans}{vector of odd integers giving the widths of modified Daniell
smoothers to be used to smooth the periodogram}

\item{B}{number of bootstrap samples.}

\item{...}{other arguments passed to [stats::spec.pgram()]}
}
\description{
A light wrapper around [stats::spec.pgram()] and [stats::ar()]. Returns the spectrum of
the signal and the spectrum of the fitted autoregressiv model.
}
