% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eof.R
\name{flip_ceof}
\alias{flip_ceof}
\title{Flip sign of eofs}
\usage{
flip_ceof(eof, n = NULL)
}
\arguments{
\item{eof}{eof object returned from [metR::EOF()]}

\item{n}{vector of eofs to flip. If `NULL`, defaults to all.}
}
\description{
Flip sign of eofs
}
