% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{geom_qmap}
\alias{geom_qmap}
\title{Quick maps}
\usage{
geom_qmap(
  subset = identity,
  crop = NULL,
  color = "gray50",
  size = 0.3,
  fill = NA,
  ...
)
}
\arguments{
\item{subset}{a function that will be applied to the map data.}

\item{crop}{optional bounding box to crop the map.}

\item{color, size, fill}{parameters passed to [ggplot2::geom_polygon()]}

\item{...}{other arguments passed to [ggplot2::geom_polygon()]}
}
\description{
Quick maps
}
