% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-tstudent.R
\name{geom_tstudent}
\alias{geom_tstudent}
\title{Draws a violin plot representing a T-Student distribution}
\usage{
geom_tstudent(
  mapping = NULL,
  data = NULL,
  stat = StatTStudent,
  position = "dodge",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  orientation = NA,
  df = 30,
  ...
)
}
\description{
Draws a violin plot representing a T-Student distribution
}
