% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lite.R
\name{lm_lite}
\alias{lm_lite}
\title{Fast linear regression}
\usage{
lm_lite(y, x, weights = rep(1, length(y)), r2 = FALSE)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variable}

\item{weights}{vector of weights}

\item{r2}{logical indicating wether to compute r.quared}
}
\description{
Computes regression coefficient quickly for the case
of simple lienar regression using the simple
covariance/variance formula.
}
