% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise-coords.R
\name{normalise_coords}
\alias{normalise_coords}
\title{Change dimension names}
\usage{
normalise_coords(
  data,
  rules = list(lev = c("level"), lat = c("latitude"), lon = c("longitude", "long"),
    time = c("date")),
  extra = list()
)
}
\arguments{
\item{data}{A data.table.}

\item{rules}{A named list defining renaming rules.}

\item{extra}{Extra rules that will be concatenates with `rules`.
This allows to quickly add extra rules to the default ones.}
}
\description{
Changes the names of columns in a dataframe according to a set of rules
}
