% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_unique.R
\name{on_unique}
\alias{on_unique}
\title{Applies a function to a vector only once for each value}
\usage{
on_unique(x, f, ...)
}
\arguments{
\item{x, }{vector}

\item{f, }{function to apply. It needs to be a function that takes 1 value and
returns 1 value.}

\item{..., }{aditional arguments passed to `f`}
}
\value{
a vector the same length as `x`
}
\description{
In cases when you have a large vector with a lot of repreated values
(for example, the dimension columns of a spatial field in a  data.table)
applying a function to each unique value of a vector and then repeating
the result is faster and more memory efficient than applying the function
to each element.
}
