% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_cor.R
\name{partial_cor}
\alias{partial_cor}
\title{Partial correlation between x1 and y and x2 and y.}
\usage{
partial_cor(y, x1, x2)
}
\arguments{
\item{y}{variable to which partial correlation will be computed}

\item{x1, x2}{variables that will be correlated with y}
}
\description{
Partial correlation between x1 and y and x2 and y.
}
