% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lite.R
\name{Pvaluate}
\alias{Pvaluate}
\title{Compute p-values}
\usage{
Pvaluate(estimate, std.error, df, adjustment = "none")
}
\arguments{
\item{estimate}{estimate of the regression (or other fit)}

\item{std.error}{standard error}

\item{df}{degrees of freedom}

\item{adjustment}{method for adjustment, see [stats::p.adjust()].}
}
\description{
Computes p-values from the result of [metR::FitLm()] (or any assumed student-t-distributed
statistic) with optional adjustments.
}
