% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lite.R
\name{report_p}
\alias{report_p}
\title{Report p-values}
\usage{
report_p(p)
}
\arguments{
\item{p}{numeric vector of p-values}
}
\value{
A character vector of the same length as `p` with `" < 0.001"` for values
lower than 0.001 and `" = signif(p, 2)"` otherwise.
}
\description{
Creates a "pretty" string from p-values. Reports p-values (adding an equal sign)
rounded to two  significant if they are greater than 0.001, in which case,
it reports " < 0.001".
}
