% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lite.R
\name{rm_intercept}
\alias{rm_intercept}
\title{Remove intercept}
\usage{
rm_intercept(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Removes intercept term from the result of [metR::FitLm()]. More to the point,
it removes rows from a data.frame in which a column named "term" equals "(Intercept)".
}
