% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lite.R
\name{rm_singleton}
\alias{rm_singleton}
\title{Remove singleton dimensions}
\usage{
rm_singleton(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
Removes any column from a dataframe that holds only
1 unique value.
}
