% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shceof-package.R
\docType{package}
\name{shceof-package}
\alias{shceof}
\alias{shceof-package}
\title{shceof: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
