% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-poisson.R
\name{solve_poisson}
\alias{solve_poisson}
\title{Solve the poisson equation}
\usage{
solve_poisson(forcing, lon, lat)
}
\arguments{
\item{forcing}{numeric vector with the forcing (the vorticity)}

\item{lon, lat}{numeric vectors with the locations of the forcing.}
}
\description{
Solve the poisson equation
}
