% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_shceof}
\alias{theme_shceof}
\title{Main theme used in the paper}
\usage{
theme_shceof(base_size = 16)
}
\arguments{
\item{base_size}{base font size, given in pts}
}
\description{
Based on [ggplot2::theme_minimal()] but with minimal (wink) modifications.
}
