% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFlux.R
\name{WaveFlux}
\alias{WaveFlux}
\title{Computes wave flux}
\usage{
WaveFlux(psi, p = 250, a = 6371000)
}
\arguments{
\item{psi}{data.table with columns lon, lat and psi.z}

\item{p}{pressure level in hPa}

\item{a}{radius of the Earth in meters}
}
\description{
Computes wave flux
}
