% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_locations.R
\name{data_path}
\alias{data_path}
\alias{ERA5}
\alias{ERA5_BE}
\alias{ERA5_SST}
\alias{ERA20C}
\alias{NCEP_PSI}
\alias{NCEP_VOR}
\alias{HADSST}
\alias{CMAP}
\alias{O3}
\title{Data locations}
\usage{
data_path(type = c("raw", "derived"), ...)

ERA5(force_download = FALSE)

ERA5_BE(force_download = FALSE)

ERA5_SST(force_download = FALSE)

ERA20C(force_download = FALSE)

NCEP_PSI(force_download = FALSE)

NCEP_VOR(force_download = FALSE)

HADSST(force_download = FALSE)

CMAP(force_download = FALSE)

O3(force_download = FALSE)
}
\arguments{
\item{type}{character indicating the type of data. The resulting base will be type_data}

\item{...}{characters to be contactenadted with [file.path()]}

\item{force_download}{Logical indicating wether to force download of data.}
}
\description{
Functions that return the location of different datasets. They check that
relevant file exist and then return it with a class that has a neat print method.
}
