% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReIm.R
\name{ReIm}
\alias{ReIm}
\alias{sep_ReIm}
\title{Splits complex numbers into real and imaginary vectors}
\usage{
ReIm(complex)

sep_ReIm(data, column, format = c("longer", "wider"))
}
\arguments{
\item{complex}{vector of complex values}

\item{data}{a data.table}

\item{column}{unquoted name of the column to split into real and imaginary part}

\item{format}{Whether to return a long or wide dataset.}
}
\description{
Splits complex numbers into real and imaginary vectors
}
